requirements.txt
Django
opencv-python
django-cors-headers
djangorestframework